restarts celery on deploy (#5653)
* restarts celery on deploy

* celery

* Updated deploy.bash
 - added a check for the celery environment

* use single brackets instead of double

* remove reliance on the z check to ensure that if check passes correctly on all hosts

* Update deploy.bash

fix the restart command to include the .service designation of celery

+++ scripts/deploy.bash
@@ -82,12 +82,16 @@ fi
 
 # let gunicorn know its ok to restart
 if ! [ "$JOBS_NODE" ]; then
-    echo "- gunicorn"
-    for pid in $(pgrep -fl "gunicorn: worke" | awk '{print $1}'); do
-    sudo kill -1 $pid
-    sleep 0.5
-    done
-
+    if ! [ "$CELERY_NODE"  ]; then
+      echo "- gunicorn"
+      for pid in $(pgrep -fl "gunicorn: worke" | awk '{print $1}'); do
+      sudo kill -1 $pid
+      sleep 0.5
+      done
+    else
+      echo "- celery"
+      sudo systemctl restart celery.service
+    fi
 fi
 
 # invalidate cloudfront
